# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Videogames API",
        "description": "API REST para gestión de videojuegos y plataformas con autenticación OAuth 2.0",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/videojuegos": {
            "get": {
                "tags": [
                    "Videojuegos"
                ],
                "summary": "Listar todos los videojuegos",
                "description": "Obtiene la lista completa de videojuegos con sus plataformas",
                "operationId": "18b356275db7186c9b3b8c2611b21bd7",
                "responses": {
                    "200": {
                        "description": "Lista de videojuegos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "The Legend of Zelda: Breath of the Wild"
                                                    },
                                                    "anio_lanzamiento": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2017-03-03"
                                                    },
                                                    "genero": {
                                                        "type": "string",
                                                        "example": "Aventura"
                                                    },
                                                    "plataformas": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 3
                                                                },
                                                                "nombre": {
                                                                    "type": "string",
                                                                    "example": "Nintendo Switch"
                                                                },
                                                                "fabricante": {
                                                                    "type": "string",
                                                                    "example": "Nintendo"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Videojuegos"
                ],
                "summary": "Crear videojuego (Solo Admin)",
                "description": "Crea un nuevo videojuego. Requiere permisos de administrador",
                "operationId": "6013d956fd639375583dce2b05d14fb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "anio_lanzamiento",
                                    "genero",
                                    "plataformas"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Super Mario Odyssey"
                                    },
                                    "anio_lanzamiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2017-10-27"
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "Plataformas"
                                    },
                                    "plataformas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Videojuego creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Videojuego creado exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "titulo": {
                                                    "type": "string"
                                                },
                                                "anio_lanzamiento": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "genero": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No tienes permisos (solo admin)"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/videojuegos/{id}": {
            "get": {
                "tags": [
                    "Videojuegos"
                ],
                "summary": "Ver videojuego por ID",
                "description": "Obtiene un videojuego específico con sus plataformas",
                "operationId": "c97467651dc03f3284acb8ec23d55ca4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del videojuego",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Videojuego encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "titulo": {
                                                    "type": "string",
                                                    "example": "Elden Ring"
                                                },
                                                "anio_lanzamiento": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2022-02-25"
                                                },
                                                "genero": {
                                                    "type": "string",
                                                    "example": "RPG"
                                                },
                                                "plataformas": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "nombre": {
                                                                "type": "string"
                                                            },
                                                            "fabricante": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Videojuego no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Videojuegos"
                ],
                "summary": "Actualizar videojuego (Solo Admin)",
                "description": "Actualiza un videojuego existente. Requiere permisos de administrador",
                "operationId": "c0b8a01ad6270804bf4e302dbbf3382b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del videojuego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Super Mario Odyssey - Edición Deluxe"
                                    },
                                    "anio_lanzamiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2017-10-27"
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "Aventura/Plataformas"
                                    },
                                    "plataformas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Videojuego actualizado exitosamente"
                    },
                    "403": {
                        "description": "No tienes permisos"
                    },
                    "404": {
                        "description": "Videojuego no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Videojuegos"
                ],
                "summary": "Eliminar videojuego (Solo Admin)",
                "description": "Elimina un videojuego. Requiere permisos de administrador",
                "operationId": "842e6a645edda0846fa27f473b2f6895",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del videojuego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Videojuego eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Videojuego eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No tienes permisos"
                    },
                    "404": {
                        "description": "Videojuego no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/videojuegos/recientes": {
            "get": {
                "tags": [
                    "Videojuegos"
                ],
                "summary": "Obtener videojuegos recientes",
                "description": "Obtiene los 3 videojuegos más recientes agregados al sistema",
                "operationId": "4a69ca62a23d61a59923faab4c8b03a6",
                "responses": {
                    "200": {
                        "description": "Videojuegos recientes obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "titulo": {
                                                        "type": "string"
                                                    },
                                                    "anio_lanzamiento": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "genero": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token Bearer obtenido del endpoint de login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}